+++ src/Nodes/CommonNode.tsx
@@ -19,8 +19,7 @@ const CommonNode: React.FC = () => {
 
   const Component = registerNode?.displayComponent || DefaultNode;
 
-  const handleNodeClick = async (e: React.MouseEvent) => {
-    e.stopPropagation();
+  const handleNodeClick = async () => {
     try {
       await beforeNodeClick?.(node);
       clickNode();
@@ -32,14 +31,15 @@ const CommonNode: React.FC = () => {
   return (
     <div className={`flow-builder-node ${registerNode?.className || ''}`}>
       <Arrow />
-      <div className="flow-builder-node__content" onClick={handleNodeClick}>
-        <Component
-          readonly={readonly}
-          node={node}
-          nodes={nodes}
-          remove={removeNode}
-        />
-
+      <div className="flow-builder-node__content-wrap">
+        <div className="flow-builder-node__content" onClick={handleNodeClick}>
+          <Component
+            readonly={readonly}
+            node={node}
+            nodes={nodes}
+            remove={removeNode}
+          />
+        </div>
         <RemoveButton />
       </div>