+++ src/Nodes/ConditionNode.tsx
@@ -55,8 +55,7 @@ const ConditionNode: React.FC<IProps> = (props) => {
     [sortableAnchor],
   );
 
-  const handleNodeClick = async (e: React.MouseEvent) => {
-    e.stopPropagation();
+  const handleNodeClick = async () => {
     try {
       await beforeNodeClick?.(node);
       clickNode();
@@ -101,15 +100,17 @@ const ConditionNode: React.FC<IProps> = (props) => {
 
       <Arrow />
 
-      <div className="flow-builder-node__content" onClick={handleNodeClick}>
-        {sortable ? <ConditionDragHandle /> : null}
-        <Component
-          readonly={readonly}
-          node={node}
-          nodes={nodes}
-          remove={removeNode}
-        />
+      <div className="flow-builder-node__content-wrap">
+        <div className="flow-builder-node__content" onClick={handleNodeClick}>
+          <Component
+            readonly={readonly}
+            node={node}
+            nodes={nodes}
+            remove={removeNode}
+          />
+        </div>
         <RemoveButton />
+        {sortable ? <ConditionDragHandle /> : null}
       </div>
 
       <AddButton />